Ticker Addition
Added Tickers to the results and adjusted accordingly.

Also made the results a little sexier by capitalising some of the
letters.

Still haven't done any major testing, but everything should be working
as intended.

Next step will be to get rid of annoying errors and maybe turning this
whole thing into an .exe.