Location Adjusting
Changed the file location again. Finally synced with github.

Gave up on removing red warning messages, they're not that bad anyways.

Also this location adjusting issue is kind of annoying and should be
fixed sometime in the future.

Currently testing, if everything is OK I will probably update only the
.exe version sometime later.