Automatization
Moved everything into a single R script AAS.R (Automatic Article
Searcher, duh) and added .bat files for a simple doubleclick execution.
Everything works as intended, already tested.

Added a readme explaining the functions and usage of files with a quick
tutorial

Also added an excel file to test if any data is an anomaly and/or if
it's just an error in the code.

If everything is working as intended this will be the last addition to
AAS. For further analysis I'll probably be switching to different
programming language.