Big Update
I will describe every change in order, by file.

- AAS.R changes:
Switched havingInternet function to a much simpler solution of testing
if google is loading. The version before wasn't always working and this
is working perfectly so far.

Added the internet connection test to keywordSearch function in order to
avoid stopping if internet crashes mid-execution. (Instead of testing
only at the start of execution)

Also, length(results) function wasn't working with the new R/Rstudio
version so I switched to nrow/ncols. (Why didn't I do this sooned)

Added try() to getURL since sometimes URL's tend to be broken and/or
limit the connectivity to less than 1/2 mins.

Minor fix in syntax.

Added an if statement to handle errors.

Some other minor fixes and adjustments.

Also added an optional code to send emails to myself with the gathered
data and a quick info on how the task was running.

Also switched up some of the data types for easier email viewing and
handling. Will do another update soon.